% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{boxPlot}
\alias{boxPlot}
\title{WQ boxplot.}
\usage{
boxPlot(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A boxplot
}
\description{
This function produces boxplots for WQ datadepending on AU_ID, WQ constituent
, or both.
}
