% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotlog10.R
\name{boxPlotlog10}
\alias{boxPlotlog10}
\title{WQ boxplot (log10 scale).}
\usage{
boxPlotlog10(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A boxplot (log10 scale)
}
\description{
This function produces boxplots (log10 scale) for WQ datadepending on AU_ID
, WQ constituent, or both.
}
