% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_AU.R
\name{categorize_AU}
\alias{categorize_AU}
\title{Categorize Overall Assessment Unit}
\usage{
categorize_AU(input_categorized_uses)
}
\arguments{
\item{input_categorized_uses}{Final results from categorize_AU_uses function.}
}
\value{
Overall category for each AU
}
\description{
The categorize_AU function assigns each AU to an IR category.
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_MagDurFreq_Results <- read_csv(system.file("extdata/AK_Final_MagDurFreq_Output.csv"
                                       , package = "AKDECtools")
                                       , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

uses_categorized <- categorize_AU_uses(df_MagDurFreq_Results, simplify_standards = F)

categorize_AU(uses_categorized)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
