% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_AU_uses.R
\name{categorize_AU_uses}
\alias{categorize_AU_uses}
\title{Categorize Assessment Units (AUs) by Use}
\usage{
categorize_AU_uses(MagDurFreq_Results, assessments, simplify_standards)
}
\arguments{
\item{MagDurFreq_Results}{Final results from MagDurFreq_combine function.}

\item{assessments}{Assessment data from ATTAINS expert query found at https://owapps.epa.gov/expertquery/attains/}

\item{simplify_standards}{True/False input to condense final category for constituents
with multiple standards per use.}
}
\value{
Categorized AUs for each use
}
\description{
The categorize_AU function assigns each AU/characteristic combination to an
IR category and then assigns an overall designated use category to the AU
for every applicable use.
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_MagDurFreq_Results <- read_csv(system.file("extdata/AK_Final_MagDurFreq_Output.csv"
                                       , package = "AKDECtools")
                                       , guess_max = 10^6)

df_Assessments <- read_csv(system.file("extdata/AK_Assessments.csv"
                                       , package = "AKDECtools")
                                       , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

categorize_AU_uses(df_MagDurFreq_Results, df_Assessments, simplify_standards = F)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
