% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCat3samples.R
\name{filterCat3samples}
\alias{filterCat3samples}
\title{Filter out IR Category 3 samples}
\usage{
filterCat3samples(data_samples, data_sufficiency)
}
\arguments{
\item{data_samples}{Water quality data in long format. Output from data_processing.R.}

\item{data_sufficiency}{Data sufficiency by AU/characteristic. Output from data_processing.R.}
}
\value{
Filtered WQ dataset not including Category 3 AU/characteristic combinations.
}
\description{
The filterCat3samples function filters a data frame of water quality samples
based on the data sufficiency output produced by data_processing.R. Only
AU/characteristic combinations with Data_Sufficient == 'Yes' are kept.
}
\details{
Required fields for input arguments
\itemize{
\item data_samples:
\item data_sufficiency:
}
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_Data_Sufficiency <- read_csv(system.file("extdata/AK_Data_Sufficiency.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

filterCat3samples(data_samples = df_ExampSamps
, data_sufficiency = df_Data_Sufficiency)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
}
