% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTurbidityReferenceSites.R
\name{findTurbidityReferenceSites}
\alias{findTurbidityReferenceSites}
\title{Identify the monitoring locations within AUs with sufficient turbidity}
\usage{
findTurbidityReferenceSites(input_samples_filtered)
}
\arguments{
\item{input_samples_filtered}{Water quality samples limited to those AUs with
sufficient data for a given characteristic. Use filterCat3samples function to filter
before running this function.}
}
\value{
Table with a column of AUs with sufficient turbidity, a column of
the monitoring location sites within that AU, and a column totalling the
number of monitoring sites within that AU.
}
\description{
This function identifies which AUs have sufficient turbidity and returns a
detailed list of monitoring locations within that AU that can be selected
as a reference site for turbidity analysis. This function is not required, but
should be run before MagDurFreq_turbidity if chosen.
}
\details{
Required fields for input arguments
\itemize{
\item input_samples_filtered: TADA.CharacteristicName, AUID_ATTNS,
MonitoringLocationIdentifier
}
}
\examples{
#' # Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)
library(dplyr)

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

list_of_needed_sites <- findTurbidityReferenceSites(df_ExampSamps)
}
}
