% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshwaterAmmoniaWarning.R
\name{freshwaterAmmoniaWarning}
\alias{freshwaterAmmoniaWarning}
\title{Freshwater ammonia warning}
\usage{
freshwaterAmmoniaWarning(input_sufficiency)
}
\arguments{
\item{input_sufficiency}{Data sufficiency table generated using the data_processsing.R script}
}
\value{
Returns message
}
\description{
This function checks if there is sufficient freshwater ammonia data for
water quality analyses. Returns message.
}
\details{
Required fields for input arguments
\itemize{
\item input_sufficiency:
}
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_Data_Sufficiency <- read_csv(system.file("extdata/AK_Data_Sufficiency.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

freshwaterAmmoniaWarning(df_Data_Sufficiency)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
