% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagDurFreq.R
\name{MagDurFreq}
\alias{MagDurFreq}
\title{Assess magnitude, duration, and frequency}
\usage{
MagDurFreq(wqs_crosswalk, input_samples_filtered, input_sufficiency)
}
\arguments{
\item{wqs_crosswalk}{Water quality standards crosswalk table}

\item{input_samples_filtered}{Water quality samples limited to those AUs with
sufficient data for a given characteristic. Use filterCat3samples function to filter
before running this function.}

\item{input_sufficiency}{Data sufficiency table generated using the data_processsing.R script}
}
\value{
A dataset with MagDurFreq results
}
\description{
The MagDurFreq function applies Alaska's surface water quality standards (SWQS)
to water quality data. Ideally, data are filtered before running this function
to only include data sufficient for water quality assessments. Water quality
exceedances are determined using the magnitude, duration, and frequency
components of Alaska's SWQS.
}
\details{
Required fields for input arguments
\itemize{
\item wqs_crosswalk: Constituent, Directionality, Frequency, Duration, Details, Use,
Waterbody Type, Magnitude_Text, Magnitude_Numeric
\item input_samples_filtered: AUID_ATTNS, ActivityStartDate, AU_Type, ActivityStartTime.Time,
TADA.ResultMeasureValue, TADA.ActivityDepthHeightMeasure.MeasureValue
\item input_sufficiency: AUID_ATTNS, TADA.CharacteristicName, Use, Waterbody Type,
}
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_WQS_Crosswalk <- read_csv(system.file("extdata/AK_WQS_Crosswalk.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_Data_Sufficiency <- read_csv(system.file("extdata/AK_Data_Sufficiency.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1
# Filter samples to only those with sufficient data to make MagDurFreq
## run more quickly

input_samples_filtered <- filterCat3samples(data_samples = df_ExampSamps
, data_sufficiency = df_Data_Sufficiency)

MagDurFreq(df_WQS_Crosswalk, input_samples_filtered, df_Data_Sufficiency)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
