% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagDurFreq_hardness.R
\name{MagDurFreq_hardness}
\alias{MagDurFreq_hardness}
\title{Assess magnitude, duration, and frequency (hardness dependent)}
\usage{
MagDurFreq_hardness(
  wqs_crosswalk,
  input_samples,
  input_samples_filtered,
  input_sufficiency
)
}
\arguments{
\item{wqs_crosswalk}{Water quality standards crosswalk table}

\item{input_samples}{All water quality samples (not limited)}

\item{input_samples_filtered}{Water quality samples limited to those AUs with
sufficient data for a given characteristic. Use filterCat3samples function to filter
before running this function.}

\item{input_sufficiency}{Data sufficiency table generated using the data_processsing.R script}
}
\value{
A dataset with MagDurFreq results
}
\description{
The MagDurFreq_hardness function applies Alaska's surface water quality
standards (SWQS) to hardness-dependent water quality data. Ideally, data are
filtered before running this function to only include data sufficient
for water quality assessments. Water quality exceedances are determined using the
magnitude, duration, and frequency components of Alaska's SWQS.
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_WQS_Crosswalk <- read_csv(system.file("extdata/AK_WQS_Crosswalk.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_Data_Sufficiency <- read_csv(system.file("extdata/AK_Data_Sufficiency.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1
# Filter samples to only those with sufficient data to make MagDurFreq
## run more quickly

input_samples_filtered <- filterCat3samples(data_samples = df_ExampSamps
, data_sufficiency = df_Data_Sufficiency)

MagDurFreq_hardness(df_WQS_Crosswalk, df_ExampSamps, input_samples_filtered
, df_Data_Sufficiency)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
