% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePull.R
\name{simplePull}
\alias{simplePull}
\title{Filter WQ data.}
\usage{
simplePull(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A filtered dataset.
}
\description{
A simple function to filter specific data from a larger WQ dataset by either AU_ID
, constituent, or both
}
