% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlotlog10.R
\name{timeSerieslog10}
\alias{timeSerieslog10}
\title{WQ time series plotting (log10 scale).}
\usage{
timeSerieslog10(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A log10-scaled plot
}
\description{
This function produces time series plots (log10 scale) for WQ data
depending on AU_ID, WQ constituent, or both.
}
