% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_AU.R
\name{categorize_AU}
\alias{categorize_AU}
\title{Categorize Assessment Units (AUs)}
\usage{
categorize_AU(MagDurFreq_Results)
}
\arguments{
\item{MagDurFreq_Results}{Final results from MagDurFreq_combine function.}
}
\value{
Categorized AUs
}
\description{
The categorize_AU function assigns each AU/characteristic combination to an
IR category and then assigns an overall designated use category to the AU
for every applicable use.
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_MagDurFreq_Results <- read_csv(system.file("extdata/AK_Final_MagDurFreq_Output.csv"
                                       , package = "AKDECtools")
                                       , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

categorize_AU(df_MagDurFreq_Results)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
