% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagDurFreq_turbidity.R
\name{MagDurFreq_turbidity}
\alias{MagDurFreq_turbidity}
\title{Assess magnitude, duration, and frequency (turbidity dependent)}
\usage{
MagDurFreq_turbidity(
  wqs_crosswalk,
  input_samples_filtered,
  input_sufficiency,
  reference_sites
)
}
\arguments{
\item{wqs_crosswalk}{Water quality standards crosswalk table}

\item{input_samples_filtered}{Water quality samples limited to those AUs with
sufficient data for a given characteristic. Use filterCat3samples function to filter
before running this function.}

\item{input_sufficiency}{Data sufficiency table generated using the data_processsing.R script}

\item{reference_sites}{List of paired reference sites for each AU in 'input_samples_filtered'}
}
\value{
A dataset with MagDurFreq results
}
\description{
The MagDurFreq_turbidity function applies Alaska's surface water quality
standards (SWQS) to turbidity-dependent water quality data. Ideally, data are
filtered before running this function to only include data sufficient
for water quality assessments. Water quality exceedances are determined using the
magnitude, duration, and frequency components of Alaska's SWQS.
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_WQS_Crosswalk <- read_csv(system.file("extdata/AK_WQS_Crosswalk.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_Data_Sufficiency <- read_csv(system.file("extdata/AK_Data_Sufficiency.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1
# Filter samples to only those with sufficient data to make MagDurFreq
## run more quickly
input_samples_filtered <- filterCat3samples(data_samples = df_ExampSamps
, data_sufficiency = df_Data_Sufficiency)

# create reference table
reference_sites <- tibble(AUID_ATTNS = au_sites
, ReferenceSites = sites)

# run MagDurFreq_hardness
MagDurFreq_hardness(df_WQS_Crosswalk, input_samples_filtered, df_Data_Sufficiency)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
