% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePull.R
\name{simplePull}
\alias{simplePull}
\title{Filter WQ data.}
\usage{
simplePull(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A filtered dataset.
}
\description{
A simple function to filter specific data from a larger WQ dataset by either AU_ID
, constituent, or both
}
\examples{
# Examples 1-4, data from Excel
\dontrun{
# Packages
library(readxl)

df_ExampSamps <- read_excel(system.file("extdata/AK_Example_Samples.xlsx"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1 - single AU and single constituent

test_pull <- simplePull(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005')
, constituent = c('PH'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 2 - multiple AUs and single constituent

test_pull <- simplePull(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005', 'AK_B_1010203_001')
, constituent = c('TEMPERATURE, WATER'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 3 - single AU and multiple constituents

test_pull <- simplePull(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005')
, constituent = c('PH', 'TEMPERATURE, WATER'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 4 - multiple AUs and multiple constituents

test_pull <- simplePull(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005', 'AK_B_1010203_001')
, constituent = c('PH', 'TEMPERATURE, WATER'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
