% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlot.R
\name{timeSeries}
\alias{timeSeries}
\title{WQ time series plotting.}
\usage{
timeSeries(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A plot
}
\description{
This function produces time series plots for WQ data
depending on AU_ID, WQ constituent, or both.
}
\examples{
# Examples 1-4, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1 - single AU and single constituent

timeSeries(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005')
, constituent = c('PH'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 2 - multiple AUs and single constituent

timeSeries(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005', 'AK_B_1010203_001')
, constituent = c('TEMPERATURE, WATER'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 3 - single AU and multiple constituents

timeSeries(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005')
, constituent = c('PH', 'TEMPERATURE, WATER'))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 4 - multiple AUs and multiple constituents

timeSeries(data = df_ExampSamps, AU_ID = c('AK_R_1010504_005', 'AK_B_1010203_001')
, constituent = c('PH', 'TEMPERATURE, WATER'))
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
