% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlot.R
\name{timeSeries}
\alias{timeSeries}
\title{WQ time series plotting.}
\usage{
timeSeries(data, WQS_table, AU_ID, y_axis_log)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{WQS_table}{Water quality standards crosswalk table}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{y_axis_log}{A TRUE/FALSE argument that specifies whether the Y-axis should be Log10 scaled}
}
\value{
A time series plot
}
\description{
The timeSeries function produces time series plots for WQ data depending on
an assessment unit identifier (AU_ID). Each WQ characteristic is plotted separately.
The y-axis can be Log10 scaled using a TRUE/FALSE argument.
}
\details{
Required fields for input arguments
\itemize{
\item data: TADA.CharacteristicName, AUID_ATTNS, ActivityStartDate,
TADA.ResultMeasureValue, MonitoringLocationIdentifier, TADA.ResultMeasure.MeasureUnitCode
\item WQS_table: TADA.Constituent
\item AU_ID: NA
\item y_axis_log: TRUE/FALSE
}
}
\examples{
# Example, data from Excel
\dontrun{
# Packages
library(readr) # readr is a tidyverse package to read CSVs
library(AKDECtools)

df_ExampSamps <- read_csv(system.file("extdata/AK_Example_Samples.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

df_WQS_Crosswalk <- read_csv(system.file("extdata/AK_WQS_Crosswalk.csv"
                                       , package = "AKDECtools")
                            , guess_max = 10^6)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Example 1

timeSeries(data = df_ExampSamps, WQS_table = df_WQS_Crosswalk
, AU_ID = c('AK_R_1010504_005'), y_axis_log = F)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}

}
