% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlot.R
\name{timeSeries}
\alias{timeSeries}
\title{A function that produces time series plots for WQ data depending on AU_ID
, WQ constituent, or both.}
\usage{
timeSeries(data, AU_ID, constituent)
}
\arguments{
\item{data}{A WQ dataset in long format. Must contain either AU_ID or TADA.CharacteristicName}

\item{AU_ID}{An AK DEC Assessment Unit (AU) Identifier}

\item{constituent}{TADA.CharacteristicName as generated after running data_processing.R}
}
\value{
A plot
}
\description{
A function that produces time series plots for WQ data depending on AU_ID
, WQ constituent, or both.
}
